rename loadbalancer resource to cleanuploadbalancers to align with others
