+++ service/controller/clusterapi/v29/resource/cleanuploadbalancers/delete.go
@@ -1,4 +1,4 @@
-package loadbalancer
+package cleanuploadbalancers
 
 import (
 	"context"
