+++ service/controller/clusterapi/v29/resource/cleanuploadbalancers/elb_test.go
@@ -1,4 +1,4 @@
-package loadbalancer
+package cleanuploadbalancers
 
 import (
 	"reflect"
