+++ service/controller/clusterapi/v29/resource/cleanuploadbalancers/error.go
@@ -1,4 +1,4 @@
-package loadbalancer
+package cleanuploadbalancers
 
 import (
 	"github.com/giantswarm/microerror"
