+++ service/controller/clusterapi/v29/resource/cleanuploadbalancers/resource.go
@@ -1,4 +1,4 @@
-package loadbalancer
+package cleanuploadbalancers
 
 import (
 	"github.com/giantswarm/microerror"
@@ -7,7 +7,7 @@ import (
 
 const (
 	// Name is the identifier of the resource.
-	Name = "loadbalancerv29"
+	Name = "cleanuploadbalancersv29"
 )
 
 // Config represents the configuration used to create a new loadbalancer resource.
